+++ ginga/misc/grc.py
@@ -37,7 +37,6 @@ def main(options, args):
     else:
         method_name = args[0]
         args, kwdargs = _grc.prep_args(args[1:])
-        #print (("args=", args, "kwdargs=", kwdargs))
 
     # invoke method on rest of parameters
     method = ginga.lookup_attr(method_name)
@@ -48,36 +47,16 @@ def main(options, args):
 
 def _main():
     """Run from command line."""
-    argprs = ArgumentParser()
+    argprs = ArgumentParser("Ginga Remote Control program")
 
-    argprs.add_argument("--debug", dest="debug", default=False,
-                        action="store_true",
-                        help="Enter the pdb debugger on main()")
     argprs.add_argument("--host", dest="host", metavar="HOST",
                         default="localhost", help="Connect to server at HOST")
-    argprs.add_argument("--port", dest="port", type=int,
-                        default=9000, metavar="PORT",
+    argprs.add_argument('-p', "--port", dest="port", type=int,
+                        default=_grc.default_rc_port, metavar="PORT",
                         help="Connect to server at PORT")
-    argprs.add_argument("--profile", dest="profile", action="store_true",
-                        default=False,
-                        help="Run the profiler on main()")
 
     (options, args) = argprs.parse_known_args(sys.argv[1:])
 
-    # Are we debugging this?
-    if options.debug:
-        import pdb
-
-        pdb.run('main(options, args)')
-
-    # Are we profiling this?
-    elif options.profile:
-        import profile
-
-        print("%s profile:" % sys.argv[0])
-        profile.run('main(options, args)')
-
-    else:
-        main(options, args)
+    main(options, args)
 
 # END